Added Random field into Dictionary
Refactor classes acording to pull request #27
