+++ extjwnl/src/main/java/net/sf/extjwnl/dictionary/Dictionary.java
@@ -34,6 +34,19 @@ import java.util.*;
 public abstract class Dictionary {
 
     private static final Logger log = LoggerFactory.getLogger(Dictionary.class);
+    
+    private Random rand = null;
+
+    public Random getRandom() {
+    	if(rand==null) {
+    		rand=new Random(new Date().getTime());
+    	}
+		return rand;
+	}
+
+	public void setRandom(Random rand) {
+		this.rand = rand;
+	}
 
     // messages for static methods
     private static final String STATIC_MESSAGES = "net.sf.extjwnl.dictionary.messages_static";
@@ -990,6 +1003,4 @@ public abstract class Dictionary {
             }
         }
     }
-
-    public abstract IndexWord getRandomIndexWord(POS pos, Random random) throws JWNLException;
 }
